Adding Traveling Salesman example; fixing minor bugs

+++ Specimen.h
@@ -65,7 +65,7 @@ namespace GA {
 		inline Genotype& getGenotype() { return dna_; }
 
 		inline double	getFitness() const { return fitness_; }
-		inline void		setFitness(int fitness) { fitness_ = fitness; }
+		inline void		setFitness(double fitness) { fitness_ = fitness; }
 	};
 
 }
